COMP: Fix -Wsign-promo in Widgets & DICOMWidgets
This commit fixes warnings like the following:

  /path/to/CTK/Libs/Widgets/Testing/Cpp/ctkCheckableHeaderViewTest1.cpp:76:73: warning: passing ‘Qt::CheckState’ chooses ‘int’ over ‘uint’ {aka ‘unsigned int’} [-Wsign-promo]
     76 |   model.setHeaderData(0, Qt::Horizontal, Qt::Checked, Qt::CheckStateRole);
        |

+++ Libs/DICOM/Widgets/Testing/Cpp/ctkDICOMBrowserTest.cpp
@@ -151,10 +151,10 @@ void ctkDICOMBrowserTester::testImportDirectoryMode()
 
   QComboBox* comboBox = browser.importDialog()->bottomWidget()->findChild<QComboBox*>();
 
-  comboBox->setCurrentIndex(comboBox->findData(ctkDICOMBrowser::ImportDirectoryCopy));
+  comboBox->setCurrentIndex(comboBox->findData(static_cast<int>(ctkDICOMBrowser::ImportDirectoryCopy)));
   QCOMPARE(browser.importDirectoryMode(), ctkDICOMBrowser::ImportDirectoryCopy);
 
-  comboBox->setCurrentIndex(comboBox->findData(ctkDICOMBrowser::ImportDirectoryAddLink));
+  comboBox->setCurrentIndex(comboBox->findData(static_cast<int>(ctkDICOMBrowser::ImportDirectoryAddLink)));
   QCOMPARE(browser.importDirectoryMode(), ctkDICOMBrowser::ImportDirectoryAddLink);
 }
